pyproject.toml
[tool.poetry]
name = "mkdocs-macros-adr-summary"
version = "0.0.0"
description = "A manager to easily handle multiple SQLAlchemy configurations"
license = "MIT"
authors = ["Federico Busetti <729029+febus982@users.noreply.github.com>"]
repository = "https://github.com/febus982/mkdocs-macros-adr-summary"
homepage = "https://febus982.github.io/mkdocs-macros-adr-summary"
readme = "README.md"
packages = [{include = "mkdocs_macros_adr_summary"}]
#keywords = ["add", "keywords", "here"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed"
]

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.dependencies]
mkdocs = "^1.4.3"
mkdocs-macros-plugin = "^1.0.5"
python = ">=3.8,<3.13"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
coverage = ">=6.5.0"
bandit = ">=1.7.6"
black = ">=22.10.0"
mkdocs-material = ">=9.1.16"
mike = ">=2.0.0"
mypy = ">=0.990"
pymdown-extensions = ">=10.0.1"
pytest = ">=7.2.0"
pytest-asyncio = ">=0.20.3"
pytest-cov = ">=4.0.0"
pytest-factoryboy = ">=2.5.0"
pytest-xdist = ">=3.0.2"
ruff = ">=0.0.263"

[tool.pytest.ini_options]
asyncio_mode = "auto"
minversion = "6.0"
addopts = "-n auto --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.mypy]
files = "mkdocs_macros_adr_summary"

[[tool.mypy.overrides]]
module = [
    "mkdocs_macros.plugin",
]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["mkdocs_macros_adr_summary"]
concurrency = ["multiprocessing"]
parallel = true

[tool.coverage.report]
exclude_also = [
    "pragma: no cover",
    "pass",
    "\\.\\.\\.",
    ]


[tool.ruff]
select = ["E", "F", "I"]
extend-exclude = ["docs"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
files = '''
(
  mkdocs_macros_adr_summary
  tests
)
'''
extend-exclude = '''
(
  /docs
)
'''
